I really want task.opts, not task.env
even though I had a test, it wasn't really working correctly!

the hint was that a dummy version looking for test/test.R was working
right, even though it should have failed because I *thought* I was
saying the working dir for R was R (which should have meant the path to
test/test.R was not a file).

modified, tested by running it, works okay.