added create_vds_wim_dist_table code
modified from the original that was in merge pairs repo.  This code can
generate distances for freeways/directions, or all to all, and return a
hash keyed on the vds id, not the wim site.  Also, while processing the
db response I clean up any extraneous WIM sites, rather than externall
to this.