hacked in a minimum vds start point setting
set the value start_vdsid in the config file under calvad, as in

{
...
   "calvad":{
        "wimpath":"/data/backup/wim/",
        "vdspath":"/data/backup/pems/breakup/",
        "outputpath":"/data/backup/pems/imputed/",
        "years":[2012],
        "truckimpute_redo":true,
        "start_vdsid":400276
    }
}

Then instead of a redo=true state forcing a redo of ALL the detectors,
instead you can just restart where the last run left off.

Needed to do this because my disk filled up with images and the process
crashed, and I didn't feel like restarting.